fix segfaults
as V8 does not clear global variables (even if V8::Dispose() is called)
clear the memory before it goes out of the main scope.

the segfaults were happening when the static data cleaner was trying to
clear the v8 pool after the main() finished and all the binding metadata
was cleared. As clearing the V8 data requires the meta class info of the
objects being cleared it was accessing an already deallocated pointer.

Solution is to clear the V8 data before the bindings are released.

Conflicts:
	lib/cpgf/samples/qtCore/sampleqtcore.cpp